+++ CHANGELOG.md
@@ -15,6 +15,7 @@ project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
   - R/clean-data.R to R/fips.R
+  - can now get clean fatalities or raw dirty fatalities.
 
 ### Removed
   - R/env-variables.R - repetitive or unnecessary code